Update xdebug.ini
XDebug is listened to by an Xdebug client (e.g. an IDE).
XDebug server (PHP) will then make a connection to the Xdebug client when the conditions are satisfied.

I used to specify `host.docker.internal` as the client host, but this was a mistake.
`xdebug.client_host = 127.0.0.1` is correct. This is because the PHP DEBUG (felixfbecker.php-debug) extension, which is installed on vscode-server, runs in a Docker container.

I've been communicating with Xdebug clients through a port that the Visual Studio Code Remote Development feature automatically port-forwarded on the Docker host machine (very inefficient!).